Reorganized the code structure
