--- schema.sql
@@ -1,17 +1,0 @@
-CREATE TABLE games (
-  id SERIAL PRIMARY KEY,
-  genre VARCHAR(255),
-  author_id INT,
-  source VARCHAR(255),
-  label VARCHAR(255),
-  publish_date DATE,
-  archived BOOLEAN,
-  multiplayer BOOLEAN,
-  last_played_at DATE
-);
-
-CREATE TABLE authors (
-  id SERIAL PRIMARY KEY,
-  first_name VARCHAR(255),
-  last_name VARCHAR(255)
-);
